requirements.txt
aiogram~=3.0.0b3
quart~=0.18.3
feedparser~=6.0.10
mysql-connector-python~=8.0.31
aiohttp~=3.8.3
vk_api